pyproject.toml
[tool.poetry]
name = "kafka-schema-registry-operator"
version = "0.1.0"
description = ""
authors = ["Daniel Middlecote <dlmiddlecote@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
kopf = "^0.28.3"
python-schema-registry-client = "^1.7.1"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"
flake8 = "^3.8.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
